pyproject.toml
[tool.poetry]
name = "10th-grade-final-project"
version = "0.1.0"
description = ""
authors = ["leterDieu <leterdieu@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.6"
psycopg2 = "^2.9.10"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
